Better way to track steps + delete confirmation

+++ src/app/components/Project.svelte
@@ -10,29 +10,38 @@
   </button>
 </div>
 
-<h3 class="text-md sm:text-lg font-bold text-coral">Row Counter</h3>
-<p class="text-sky-50">{project.rowCount}</p>
-<button
-  class="px-6 py-3 bg-sky-800 text-sky-50 rounded-lg"
-  on:click={decreaseRowCount}
->
-  -
-</button>
-<button
-  class="px-6 py-3 bg-sky-800 text-sky-50 rounded-lg"
-  on:click={increaseRowCount}
->
-  +
-</button>
-
 <h3 class="text-md sm:text-lg font-bold text-coral">Steps</h3>
 <ul class="text-sky-50">
-  {#each project.steps as step}
+  {#each project.steps as step, index}
     <li>
+      {#if project.currentStep > index}
+        ✅
+      {/if}
       {step.description}
     </li>
   {/each}
 </ul>
+
+<h3 class="text-md sm:text-lg font-bold text-coral">Current Step</h3>
+
+{#if project.currentStep >= project.steps.length}
+  <p class="text-sky-50">Done!</p>
+{:else}
+  <p class="text-sky-50">{project.steps[project.currentStep].description}</p>
+{/if}
+
+<button
+  class="px-6 py-3 bg-sky-800 text-sky-50 rounded-lg"
+  on:click={previousStep}
+>
+  Previous
+</button>
+<button
+  class="px-6 py-3 bg-sky-800 text-sky-50 rounded-lg"
+  on:click={nextStep}
+>
+  Done
+</button>
 
 <script>
   import { createEventDispatcher } from 'svelte'
@@ -42,29 +51,31 @@
   
   const dispatch = createEventDispatcher()
 
-  function increaseRowCount() {
+  function previousStep() {
     $projects.map(p => {
       if (p.id === project.id) {
-        project.rowCount += 1
-        p.rowCount = project.rowCount
+        project.currentStep = p.currentStep <= 0 ? 0 : p.currentStep - 1
+        p.currentStep = project.currentStep
       }
     })
     localStorage.setItem('projects', JSON.stringify($projects))
   }
 
-  function decreaseRowCount() {
+  function nextStep() {
     $projects.map(p => {
       if (p.id === project.id) {
-        project.rowCount = p.rowCount <= 0 ? 0 : p.rowCount - 1
-        p.rowCount = project.rowCount
+        project.currentStep = p.currentStep >= p.steps.length ? p.steps.length : p.currentStep + 1
+        p.currentStep = project.currentStep
       }
     })
     localStorage.setItem('projects', JSON.stringify($projects))
   }
 
   function deleteProject(id) {
-    $projects = $projects.filter(p => p.id !== id)
-    localStorage.setItem('projects', JSON.stringify($projects))
-    dispatch('back')
+    if (window.confirm(`Are you sure you want to delete ${project.name}?`)) {
+      $projects = $projects.filter(p => p.id !== id)
+      localStorage.setItem('projects', JSON.stringify($projects))
+      dispatch('back')
+    }
   }
 </script>
